Same benchmark with null values
It seems, that writing a null in the ctor in release
mode is not measurably slower than not writing anything.
(Possibly because the class has final fields, so there
is already a release fence anyway.)

However, writing a null in the ctor in volatile mode
is definitely slower.

Thus, if in general we can relax the volatile store in
the ctor to a release store, the we don't need separate
ctors for the "uninitialized" case, since we can just
initialize with null.